Reuse SearchResults component for favorites
